pyspark/requirements.txt
numpy
pandas
pyspark
delta-spark
kafka-python
pyex